+++ src/Bridge/Doctrine/Orm/Filter/Type/DateTimeFilterType.php
@@ -10,6 +10,7 @@ use Kreyu\Bundle\DataTableBundle\Filter\FilterData;
 use Kreyu\Bundle\DataTableBundle\Filter\FilterInterface;
 use Kreyu\Bundle\DataTableBundle\Filter\Operator;
 use Symfony\Component\Form\Extension\Core\Type\DateTimeType;
+use Symfony\Component\OptionsResolver\Options;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class DateTimeFilterType extends AbstractFilterType
@@ -29,17 +30,17 @@ class DateTimeFilterType extends AbstractFilterType
                 ],
                 'active_filter_formatter' => $this->getFormattedActiveFilterString(...),
             ])
-            ->addNormalizer('form_options', function (OptionsResolver $resolver, array $value): array {
+            ->addNormalizer('form_options', function (Options $options, array $value): array {
                 return $value + ['widget' => 'single_text'];
             })
-            ->addNormalizer('empty_data', function (OptionsResolver $resolver, string|array $value): string|array {
-                if (DateTimeType::class !== $resolver['form_type']) {
+            ->addNormalizer('empty_data', function (Options $options, string|array $value): string|array {
+                if (DateTimeType::class !== $options['form_type']) {
                     return $value;
                 }
 
                 // Note: because choice and text widgets are split into three fields under "date" index,
                 //       we have to return an array with three empty "date" values to properly set the empty data.
-                return match ($resolver['form_options']['widget'] ?? null) {
+                return match ($options['form_options']['widget'] ?? null) {
                     'choice', 'text' => [
                         'date' => ['day' => '', 'month' => '', 'year' => ''],
                     ],