+++ src/Exporter/Form/Type/ExportStrategyType.php
@@ -17,6 +17,11 @@ class ExportStrategyType extends AbstractType
             'class' => ExportStrategy::class,
             'choice_translation_domain' => 'KreyuDataTable',
             'choice_label' => 'label',
+            // TODO: Remove after removing deprecated export strategy enum cases
+            'choices' => [
+                ExportStrategy::IncludeCurrentPage,
+                ExportStrategy::IncludeAll,
+            ],
         ]);
     }